feat: gate native compiler use on env
The native compiler mentions that it is slower when large amounts of
files are used, which correlates to build times we are seeing locally.
Additionally, the native build seems to have an issue when run from Git
Bash (Cygwin and some other options worked). To address this, we will
continue to use the Java version, but leave the native compiler as an
option via an environment variable. This allows things like Docker
containers to avoid the Java dependency.